requirements.txt
# ExpressDS - Automated Data Science Platform
# Version constraints aligned with PyCaret compatibility

# Core Framework
streamlit>=1.28.0
pandas>=1.5.0,<2.2.0
numpy>=1.21.0,<2.0

# AutoML (PyCaret 3.x needs matplotlib<3.8, numpy<2)
pycaret>=3.0.0
matplotlib>=3.3.0,<3.8.0

# AI / Agentic Brain (Gemini) - REST API works without langchain-google-genai
# For full pandas agent on Python 3.10+: pip install langchain-google-genai
langchain>=0.1.0
langchain-community>=0.0.20
langchain-experimental>=0.0.50

# Data Visualization
seaborn>=0.12.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
Pillow>=10.0.0
joblib<1.4
